+++ src/components/ui/BaseSpinner.vue
@@ -1,0 +1,14 @@
+<template functional>
+  <FontAwesomeIcon
+    icon="fa-solid fa-circle-notch"
+    spin
+    style="color: rgb(207, 207, 207)"
+    size="3x"
+  />
+</template>
+
+<script>
+export default {
+  name: 'BaseSpinner',
+};
+</script>