+++ src/views/HomeView.vue
@@ -1,16 +1,23 @@
 <template>
   <div class="home-view">
-    <LandingHero />
+    <ErrorBoundary
+      componentName="LandingHero"
+      errorMessage="Unable to load landing hero"
+    >
+      <LandingHero />
+    </ErrorBoundary>
   </div>
 </template>
 
 <script>
+import ErrorBoundary from '@/components/common/ErrorBoundary.vue';
 import LandingHero from '@/components/home/LandingHero.vue';
 
 export default {
   name: 'HomeView',
   title: 'Universe Explorer',
   components: {
+    ErrorBoundary,
     LandingHero,
   },
 };
@@ -18,6 +25,6 @@ export default {
 
 <style lang="scss" scoped>
 .home-view {
-  height: 100%;
+  height: 100vh;
 }
 </style>